renamed .h files to .hpp
