+++ AcoBurd/include/my_clock.h
@@ -1,0 +1,6 @@
+#ifndef MY_CLOCK_H
+#define MY_CLOCK_H
+
+long InternalClock(void);
+
+#endif
