+++ AcoBurd/src/main.cpp
@@ -1,11 +1,12 @@
-#include "globals.hpp"
-#include "motor.hpp"
-#include "subroutines.hpp"
-#include "display.hpp"
-#include "radio.hpp"
-#include "sleep.hpp"
-#include "gps.hpp"
-#include "watchdog.hpp"
+#include "globals.h"
+#include "motor.h"
+#include "my_clock.h"
+#include "subroutines.h"
+#include "display.h"
+#include "radio.h"
+#include "sleep.h"
+#include "gps.h"
+#include "watchdog.h"
 #include "LoRa_APP.h"
 
 static RadioEvents_t RadioEvents;
