+++ AcoBurd/src/my_clock.cpp
@@ -1,4 +1,3 @@
-#include "globals.hpp"
 #include "GPS_Air530Z.h"
 
 // Global timer routine
